requirements.txt
asyncio==3.4.3
autopep8==1.6.0
azure-core==1.24.1
azure-eventhub==5.10.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
requests==2.28.0
six==1.16.0
sodapy==2.1.1
toml==0.10.2
typing_extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
